refactor(epic-38): relocate InlineToolLoopRunner into Tamma.Api
Move the agentic tool-loop core (InlineToolLoopRunner + IInlineToolLoopRunner
+ result records) out of the Tamma.Activities workflow-engine assembly and into
Tamma.Api.Services.Agents, where it is DI-registered and executed. The engine
assembly no longer hosts the API-plane LLM core.

Because the type left the analyzed engine surface, its TAMMA001 denylist entries
and the whole-type exemption in Allowlist.cs are now dead code and are removed
(a different engine type injecting it can no longer even reference it). The four
runner tests move from Tamma.Activities.Tests to Tamma.Api.Tests (the runner is
now a Tamma.Api public type); NoDirectLlmCallTests drops it from the engine-scan
allowlist since it is no longer an engine file.

Behavior unchanged. Build clean; Tamma.Activities.Tests (2086), Tamma.Api.Tests
(4035), Tamma.Activities.Guardrails.Tests (30) all green.
