+++ apps/tamma-elsa/tests/Tamma.Api.Tests/Agents/AgenticToolLoopTests.cs
@@ -8,8 +8,9 @@ using Tamma.Activities.LlmCall;
 using Tamma.Activities.LlmCall.Models;
 using Tamma.Activities.LlmCall.Tools;
 using Tamma.Activities.Security;
+using Tamma.Api.Services.Agents;
 
-namespace Tamma.Activities.Tests.LlmCall;
+namespace Tamma.Api.Tests.Agents;
 
 /// <summary>
 /// Tests for the agentic tool loop in CallLlmInlineActivity (Story 12.2).
